+++ Changelog.rst
@@ -3,8 +3,9 @@ Version NEXTVERSION
 
 **2025-??-??**
 
-* New methods: `cf.Field.to_units`, `cf.Data.to_units`
-  (https://github.com/NCAS-CMS/cf-python/issues/???)
+* New methods to allow changing units in a chain: `cf.Field.to_units`,
+  `cf.Data.to_units`
+  (https://github.com/NCAS-CMS/cf-python/issues/874)
 
 ----